pyproject.toml
[project]
name = "ci-cd-hub"
version = "1.0.0"
description = "Centralized CI/CD orchestration hub for multi-repository pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "jguida941"}
]
keywords = ["ci", "cd", "github-actions", "devops", "automation"]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.21.0",
    "defusedxml>=0.7.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mutmut>=2.4.0",
    # Linting & Formatting
    "ruff>=0.8.0",
    "yamllint>=1.33.0",
    # Type Checking
    "mypy>=1.8.0",
    "types-PyYAML>=6.0",
    "types-jsonschema>=4.0",
    # Security
    "bandit[toml]>=1.7.9",
    "pip-audit>=2.7.0",
    # Pre-commit
    "pre-commit>=3.6.0",
    # License compliance
    "pip-licenses>=4.3.0",
]
ci = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.8.0",
    "black>=24.0",
    "isort>=5.12.0",
    "mypy>=1.8.0",
    "bandit[toml]>=1.7.9",
    "pip-audit>=2.7.0",
    "mutmut>=2.4.0",
    "hypothesis>=6.0",
    "semgrep>=1.90.0",
]
wizard = [
    "questionary>=2.0.0",
    "rich>=13.0.0",
]

[project.scripts]
hub-config = "scripts.load_config:main"
hub-report = "scripts.aggregate_reports:main"
cihub = "cihub.__main__:main"

[tool.setuptools.packages.find]
include = ["cihub*"]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = ["_quarantine"]

[tool.ruff.lint]
select = ["E", "F", "W", "S", "B", "I"]
ignore = ["S101"]  # Allow assert in tests

[tool.ruff.format]
# Black-compatible formatting
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ["py310"]
extend-exclude = "_quarantine"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["cihub", "scripts"]
omit = [
    "scripts/run_aggregation.py",
    "scripts/debug_orchestrator.py",
    "scripts/check_quarantine_imports.py",
    "scripts/run_cli_integration.py",
    "scripts/validate_summary.py",
    "scripts/verify_hub_matrix_keys.py",
    "cihub/wizard/*",
    "cihub/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

# NOTE: mutmut only tests cihub/wizard/validators.py because it's the only
# module with full test coverage. The rest of cihub/ has 193 potential mutants
# with NO test coverage. To get meaningful mutation scores, add tests first!
[tool.mutmut]
paths_to_mutate = [
    "cihub/wizard/validators.py",
]
tests_dir = ["tests/"]
also_copy = ["cihub/", "scripts/", "config/", "schema/", "templates/", ".github/", "pyproject.toml"]
mutate_only_covered_lines = true
